requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
anyio==4.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
discord.py==2.3.2
frozenlist==1.4.1
fuzzywuzzy==0.18.0
google==3.0.0
google-api-core==2.17.0
google-api-python-client==2.117.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
idna==3.6
Levenshtein==0.25.0
multidict==6.0.5
oauth2client==4.1.3
oauthlib==3.2.2
protobuf==4.25.2
pyaes==1.6.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.0.1
python-Levenshtein==0.25.0
rapidfuzz==3.6.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
table2ascii==1.1.3
TgCrypto==1.2.5
uritemplate==4.1.1
urllib3==2.2.0
uvloop==0.19.0
wcwidth==0.2.13
yarl==1.9.4
